requirements.txt
opencv-python>=4.8.0
Pillow>=10.1.0
colorama>=0.4.6
numpy>=1.24.4
